restructured components and containers
